cliente-requirements.txt
fastapi==0.89.1
uvicorn[standard]==0.20.0
pulsar-client==2.10.2
pulsar-client[avro]==2.10.2
SQLAlchemy==1.4.44
requests==2.28.2